src/2-transform/requirements.txt
# ------------------------------------ app\3-transform\requirements.txt ------------------------------------ 
azure-storage-file-datalake==12.13.0
azure-identity==1.12.0
pandas==1.5.3
tenacity==8.2.3
python-dateutil==2.8.2